+++ src/djlint/__init__.py
@@ -404,6 +404,12 @@ def _fail_with_usage_code(func: Callable[..., None]) -> Callable[..., None]:
     help="Do not attempt to format set contents.",
 )
 @click.option(
+    "--quote-style",
+    type=click.Choice(("double", "single")),
+    help="Quotes to use for strings inside template tags. [default: double]",
+    show_default=False,
+)
+@click.option(
     "--max-blank-lines",
     type=int,
     help="Consolidate blank lines down to x lines. [default: 0]",
@@ -466,6 +472,7 @@ def main(
     no_line_after_yaml: bool,
     no_function_formatting: bool,
     no_set_formatting: bool,
+    quote_style: str | None,
     max_blank_lines: int | None,
     github_output: bool | None = None,
 ) -> None:
@@ -534,6 +541,7 @@ def main(
         no_line_after_yaml=no_line_after_yaml,
         no_function_formatting=no_function_formatting,
         no_set_formatting=no_set_formatting,
+        quote_style=quote_style,
         max_blank_lines=max_blank_lines,
         github_output=github_output,
         stdin="-" in src,